requirements.txt
pkgconfig==1.5.1
Flask==1.1.1
Flask-MQTT==1.1.0
paho-mqtt==1.5.0
